Collapse client brand plumbing to two brands; fold the MUI theme
The client still modeled four brand codes and three MUI themes, with
CodeaiTheme a thin palette wrapper over a paletteless CdoTheme base.
Only codeai-next and codeai-audit exist now, so:

- BrandCode shrinks to the two remaining codes in apps/src/util/brand.ts
  and both getMuiThemeForBrand copies become audit-or-default.
- The CADS purple palette moves into the base createTheme call and the
  themes/code.org directory becomes themes/codeai. CdoTheme survives as
  a deprecated alias of CodeaiTheme so existing imports compile; it goes
  away with the codeai-next -> codeai rename. CodeaiAuditTheme stays a
  one-line palette override.
- The Storybook brand toolbar, BrandDecorator, demo hosts, tests, and
  the themes README drop the legacy brands; the README's brand-scoped
  component-override convention is retired to a historical note since
  the cutover unwrapped those blocks.

Known pre-existing failure, unrelated to this change: ts-jest under
node10 module resolution cannot resolve @code-dot-org/* package
subpaths, which breaks ExperimentsPageTest/CodeWorkspaceTest the same
way it breaks untouched tests like ChangeEmailModalTest.
