+++ frontend/apps/design-system-storybook/README.md
@@ -9,7 +9,7 @@ Storybook for the `@code-dot-org/component-library` design system.
 ## Brand switcher
 
 The toolbar has a **Brand** dropdown (paintbrush icon) for previewing any
-story under each brand in the rebrand. Selecting a brand drives
+story under each brand. Selecting a brand drives
 `decorators/BrandDecorator.tsx`, which:
 
 - writes `data-brand="<code>"` onto the preview `<html>`, switching the CSS
@@ -20,12 +20,9 @@ story under each brand in the rebrand. Selecting a brand drives
 This mirrors production, where `data-brand` is set server-side on `<html>` by
 `Cdo::Brand` (see `lib/cdo/brand.rb`). The brand codes:
 
-| Code           | Toolbar label         | Tokens                                                 |
-| -------------- | --------------------- | ------------------------------------------------------ |
-| `code`         | code.org              | legacy `:root` ramp (== `codeai`)                      |
-| `codeai`       | CodeAI                | legacy `:root` ramp                                    |
-| `codeai-next`  | CodeAI · CADS         | CADS ramp (`[data-brand='codeai-next']`) — **default** |
-| `codeai-audit` | CodeAI · Audit (pink) | all-pink audit (`[data-brand='codeai-audit']`)         |
+| Code           | Toolbar label         | Tokens                                         |
+| -------------- | --------------------- | ---------------------------------------------- |
+| `codeai-next`  | CodeAI · CADS         | CADS ramp at `:root` — **default**             |
+| `codeai-audit` | CodeAI · Audit (pink) | all-pink audit (`[data-brand='codeai-audit']`) |
 
 The default is `codeai-next`, matching the production `default-brand` DCDO key.
-`code` and `codeai` render identically until the brand cutover.
