+++ frontend/packages/component-library/src/button/LinkButton.tsx
@@ -17,7 +17,7 @@ const LinkButton: React.FunctionComponent<LinkButtonProps> = props => (
  * ###  Status: ```DEPRECATED```
  *
  * @deprecated Use MUI `Button` with `href` prop from `@mui/material` instead.
- * Style overrides are in `src/themes/code.org/styleOverrides/button.tsx`.
+ * Style overrides are in `src/themes/codeai/styleOverrides/button.tsx`.
  * See `src/button/BUTTON_MIGRATION_TO_MUI.md` for migration guide.
  * Codemod available: `yarn codemod:buttons`.
  */
