+++ frontend/packages/component-library/src/themes/codeai/styleOverrides/footer.ts
