+++ frontend/packages/component-library/src/themes/createBrandTheme.ts
@@ -1,13 +1,12 @@
 import {createTheme, SimplePaletteColorOptions, Theme} from '@mui/material';
 
-import CdoTheme from './code.org';
+import CodeaiTheme from './codeai';
 
 /**
  * Builds a brand MUI theme that only overrides the primary palette,
- * inheriting all typography and component style overrides from CdoTheme via
- * deep merge. Shared by every brand whose theme differs from CdoTheme in
- * palette alone (CodeaiTheme, CodeaiAuditTheme, ...).
+ * inheriting all typography and component style overrides from CodeaiTheme
+ * via deep merge. Its only consumer is CodeaiAuditTheme.
  */
 export function createBrandTheme(primary: SimplePaletteColorOptions): Theme {
-  return createTheme(CdoTheme, {palette: {primary}});
+  return createTheme(CodeaiTheme, {palette: {primary}});
 }
