+++ frontend/packages/component-library/src/themes/index.ts
@@ -1,4 +1,3 @@
-export {default as CdoTheme} from './code.org';
-export {default as CodeaiTheme} from './codeai';
+export {default as CodeaiTheme, CdoTheme} from './codeai';
 export {default as CodeaiAuditTheme} from './codeai-audit';
 export {getMuiThemeForBrand} from './getMuiThemeForBrand';
