+++ frontend/packages/users/src/main.tsx
@@ -1,5 +1,5 @@
 // Standalone dev host. Mirrors Studio's styling foundation (fonts, color
-// tokens, CdoTheme) so the page looks like code.org outside the host.
+// tokens, CodeaiTheme) so the page matches Studio outside the host.
 import '@code-dot-org/fonts/brands/code.org/index.css';
 import '@code-dot-org/component-library-styles/fontVariables.css';
 import '@code-dot-org/component-library-styles/shapeAndSpacingVariables.css';
